pyproject.toml
[project]
name = "bark"
version = "0.1.0"
description = "A chatbot for ScottyLabs with OpenRouter and Slack integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "slack-sdk>=3.27.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    # Context Engine
    "chromadb>=0.4.0",
    "gitpython>=3.1.0",
    # Notion Integration
    "notion-client>=2.0.0",
    # Google Drive Integration
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
    "openpyxl>=3.1.0",
    # PDF OCR
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",
    "firecrawl-py>=4.14.0",
]

[project.scripts]
bark = "bark.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bark"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
